Move thunk tests into file.
